chore(sms): remove SMS / Twilio messaging integration
Deletes internal/messaging, the messaging notifier, all Twilio
config, the /webhooks/sms/inbound route, AllowedSender store
methods, and the docs/sms-setup.md doc. Postgres
allowed_senders table left in place; migration 013 (slice #5)
drops it.

Closes #3
