+++ cmd/backflow/main.go
@@ -19,7 +19,6 @@ import (
 	"github.com/backflow-labs/backflow/internal/debug"
 	"github.com/backflow-labs/backflow/internal/discord"
 	"github.com/backflow-labs/backflow/internal/embeddings"
-	"github.com/backflow-labs/backflow/internal/messaging"
 	"github.com/backflow-labs/backflow/internal/models"
 	"github.com/backflow-labs/backflow/internal/notify"
 	"github.com/backflow-labs/backflow/internal/orchestrator"
@@ -104,20 +103,6 @@ func main() {
 		log.Info().Str("url", cfg.WebhookURL).Msg("webhook notifications enabled")
 	}
 
-	// Initialize messaging
-	var messenger messaging.Messenger
-	switch cfg.SMSProvider {
-	case "twilio":
-		messenger = messaging.NewTwilioMessenger(cfg.TwilioAccountSID, cfg.TwilioAuthToken, cfg.SMSFromNumber)
-		log.Info().Str("from", cfg.SMSFromNumber).Msg("twilio SMS messaging enabled")
-	default:
-		messenger = messaging.NoopMessenger{}
-	}
-
-	if cfg.SMSProvider != "" && cfg.SMSOutboundEnabled {
-		bus.Subscribe(notify.NewMessagingNotifier(messenger, cfg.SMSEvents))
-	}
-
 	// Declare as the interface type so a nil *Uploader stays a nil interface
 	// (avoids the nil-pointer-in-non-nil-interface trap in saveTaskMetadata).
 	var s3Uploader orchestrator.S3Client
@@ -158,16 +143,6 @@ func main() {
 
 	// Debug stats endpoint (outside /api/v1/; auth is applied explicitly here).
 	router.With(api.AuthMiddleware(db, cfg.APIKey)).Get("/debug/stats", debug.StatsHandler(orch.Running, db, startedAt).ServeHTTP)
-
-	// Mount SMS inbound webhook only when both provider and auth token are configured.
-	// The auth token is required for Twilio signature validation — without it the
-	// endpoint would accept unauthenticated requests.
-	if cfg.SMSProvider != "" && cfg.TwilioAuthToken != "" {
-		router.Post("/webhooks/sms/inbound", messaging.InboundHandler(db, cfg, messenger))
-		log.Info().Msg("SMS inbound webhook mounted at /webhooks/sms/inbound")
-	} else if cfg.SMSProvider != "" {
-		log.Warn().Msg("SMS inbound webhook NOT mounted: TWILIO_AUTH_TOKEN is required")
-	}
 
 	// Discord integration
 	if cfg.DiscordEnabled() {
